requirements.txt
translate
pytesseract
pillow